[Refactor] Auth, Tracker Swagger Docs 수정
